% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db-lib.R
\name{askDataDirEnvironment}
\alias{askDataDirEnvironment}
\title{setDataDirEnv}
\usage{
askDataDirEnvironment(env = "PACKAGE")
}
\arguments{
\item{env}{The type of environment to work with. Values are "PACKAGE" or "HOME" and it defaults to package}
}
\description{
Sets the path of Rpolyhedra data dir.
}
\details{
This function is used to set the data directories either to the package or the user home directory.
}
