% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db-lib.R
\name{checkDatabaseVersion}
\alias{checkDatabaseVersion}
\title{checkDatabaseVersion}
\usage{
checkDatabaseVersion()
}
\value{
"UPDATE" if an update is required, "NO_ACTION_REQUIRED"
     otherwise.
}
\description{
Determines if there is a need for a database update by checking the
version file of both the package and the database installation.
}
