% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polyhedra-lib.R
\name{checkVertices}
\alias{checkVertices}
\title{checkVertices()}
\usage{
checkVertices(vertices, transformed.vertices, triangulated.solid)
}
\arguments{
\item{vertices}{vertices dataframe for checking}

\item{transformed.vertices}{positioned vertices dataframe for checking}

\item{triangulated.solid}{triangulated.solid for checking}
}
\value{
checked positioned vertices
}
\description{
Check rendering vertices properties
}
