% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db-lib.R
\name{compatiblePolyhedraRDS}
\alias{compatiblePolyhedraRDS}
\title{compatiblePolyhedraRDS()}
\usage{
compatiblePolyhedraRDS(.polyhedra = .polyhedra)
}
\arguments{
\item{.polyhedra}{current polyhedra database}
}
\description{
Tests if the polyhedra RDS is compatible with the current format
}
