% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml-lib.R
\name{convertDFToXML}
\alias{convertDFToXML}
\title{convertDFToXML()}
\usage{
convertDFToXML(df, name, node)
}
\arguments{
\item{df}{the dataframe to convert}

\item{name}{the tag for the new section}

\item{node}{the node that will be acting as parent.}
}
\value{
the node with the dataframe in it
}
\description{
Allows to the creation of xml excerpts out of a data.frame.
}
