% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-lib.R
\name{downloadRPolyhedraSupportingFiles}
\alias{downloadRPolyhedraSupportingFiles}
\title{downloadRPolyhedraSupportingFiles}
\usage{
downloadRPolyhedraSupportingFiles()
}
\value{
TRUE if sucessfull, FALSE otherwise
}
\description{
Downloads the files from the remote location
}
