% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/env-lib.R
\name{genLogger}
\alias{genLogger}
\title{genLogger}
\usage{
genLogger(r6.object)
}
\arguments{
\item{r6.object}{an r6.object}
}
\description{
Returns a configured logger with threshold according r6 object.
This function is usually called in class constructors
}
\author{
ken4rab
}
