% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polyhedra-lib.R
\name{getAvailablePolyhedra}
\alias{getAvailablePolyhedra}
\title{Gets the list of available polyhedra}
\usage{
getAvailablePolyhedra()
}
\value{
polyhedra names vector
}
\description{
Gets the list of available polyhedra
}
\examples{
getAvailablePolyhedra()

}
