% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db-lib.R
\name{getAvailableSources}
\alias{getAvailableSources}
\title{getAvailableSources()}
\usage{
getAvailableSources()
}
\value{
sources string vector
}
\description{
Gets the list of names of available sources in database
}
\examples{
#gets all sources in the database
available.sources <- getAvailableSources()

#returns all polyhedra from all sources
available.polyhedra <- getAvailablePolyhedra(sources=available.sources)

#search within the polyhedron names from all sources
cubes <- getAvailablePolyhedra(sources=available.sources,
        search.string="cube")
}
