% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polyhedra-lib.R
\name{getDataDir}
\alias{getDataDir}
\title{Get the path of package data}
\usage{
getDataDir()
}
\description{
Get the path of package data
}
\examples{
## returns current data directory for Rpolyhedra lib
getDataDir()
}
