% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db-lib.R
\name{getEnvironmentFilepath}
\alias{getEnvironmentFilepath}
\title{getEnvironmentFilepath}
\usage{
getEnvironmentFilepath()
}
\value{
The environment filepath
}
\description{
Gets the filename where package data environment is persisted
}
