% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db-lib.R
\name{getGitCommit}
\alias{getGitCommit}
\title{getGitCommit
get the last git commit sha}
\usage{
getGitCommit(long.version = FALSE)
}
\arguments{
\item{long.version}{determines if the complete version of the sha will
be returned.}
}
\value{
String with git commit sha
}
\description{
getGitCommit
get the last git commit sha
}
