% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/env-lib.R
\name{getLogger}
\alias{getLogger}
\title{getLogger}
\usage{
getLogger(r6.object)
}
\arguments{
\item{r6.object}{an r6.object}
}
\description{
Returns the configured lgr of an r6 object.
If the object don't have a lgr or is not initialized returns an error
}
\author{
ken4rab
}
