% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db-lib.R
\name{getPercentilPolyhedraQuant}
\alias{getPercentilPolyhedraQuant}
\title{getPercentilPolyhedraQuant()
returns polyhedra quantity of parameter percentil}
\usage{
getPercentilPolyhedraQuant(percentil)
}
\description{
getPercentilPolyhedraQuant()
returns polyhedra quantity of parameter percentil
}
