% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/public-lib.R
\name{getPolyhedraObject}
\alias{getPolyhedraObject}
\title{Get a polyhedra object}
\usage{
getPolyhedraObject()
}
\value{
.polyhedra
}
\description{
Return the polyhedra database handler.
}
\seealso{
PolyhedraDatabase
}
