% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polyhedra-lib.R
\name{getPolyhedraRDSPath}
\alias{getPolyhedraRDSPath}
\title{Get the path of Polyhedra RDS file}
\usage{
getPolyhedraRDSPath(polyhedra.rds.filename = "polyhedra.RDS")
}
\arguments{
\item{polyhedra.rds.filename}{filename of polyhedra database}
}
\description{
Get the path of Polyhedra RDS file
}
