% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polyhedra-lib.R
\name{getPolyhedron}
\alias{getPolyhedron}
\title{Gets a polyhedron from the database.}
\usage{
getPolyhedron(name)
}
\arguments{
\item{name}{a valid name of a polyhedron in the database}
}
\description{
Gets a polyhedron from the database.
}
