% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db-lib.R
\name{getPreloadedDataFilename}
\alias{getPreloadedDataFilename}
\title{getPreloadedDataFilename}
\usage{

  getPreloadedDataFilename(polyhedra_preloaded_data = "polyhedra.preloaded.data.csv")
}
\arguments{
\item{polyhedra_rds_filename}{filename of polyhedra database}
}
\value{
the path to the Polyhedra database file
}
\description{
Gets the path of Polyhedra preloaded data CSV file
}
