% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db-lib.R
\name{getUserSpace}
\alias{getUserSpace}
\title{getUserSpace}
\usage{
getUserSpace()
}
\arguments{
\item{data.env}{enviroment where data directory must be returned}
}
\value{
path of user space
}
\description{
This function is used internally for accesing the local database path
}
