% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db-lib.R
\name{isCompatiblePolyhedraRDS}
\alias{isCompatiblePolyhedraRDS}
\title{isCompatiblePolyhedraRDS()}
\usage{
isCompatiblePolyhedraRDS(.polyhedra.candidate = .polyhedra)
}
\arguments{
\item{.polyhedra.candidate}{current polyhedra database candidate for checking compatibility}
}
\description{
Tests if the polyhedra RDS is compatible with the current format
}
