% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/env-lib.R
\name{loggerSetupFile}
\alias{loggerSetupFile}
\title{loggerSetupFile}
\usage{
loggerSetupFile(log.file, default.threshold = "info", append = TRUE)
}
\arguments{
\item{log.file}{log path for logging file}

\item{default.threshold}{threshold for setting root. Default = "info"}

\item{append}{if set to FALSE, cleanup all previous logs}
}
\description{
loggerSetupFile
}
\author{
kenarab
}
