% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ledger-lib.R
\name{maxWithoutNA}
\alias{maxWithoutNA}
\title{maxWithoutNA
Function that returns NA if all elements are NA, and the max value not NA, if not.}
\usage{
maxWithoutNA(x)
}
\arguments{
\item{x}{vector for processing function}
}
\description{
maxWithoutNA
Function that returns NA if all elements are NA, and the max value not NA, if not.
}
