% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/env-lib.R
\name{mutate_cond}
\alias{mutate_cond}
\title{mutate_cond}
\usage{
mutate_cond(.data, condition, ..., envir = parent.frame())
}
\arguments{
\item{.data}{data frame to apply the mutate}

\item{condition}{condition to conditionally apply mutate}

\item{...}{mutation function}

\item{envir}{environment to apply condition}
}
\description{
mutate_cond
}
