% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serialization-lib.R
\name{persistPolyhedron}
\alias{persistPolyhedron}
\title{persistPolyhedron}
\usage{
persistPolyhedron(polyhedron.state.defined, file.path)
}
\arguments{
\item{polyhedron.state.defined}{the polyhedron to get a representation from}

\item{file.path}{the path of the file to be persisted.}
}
\value{
the path or None
}
\description{
persists a polyhedron to the defined path in a Zipped ascii RDS format
}
