% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db-lib.R
\docType{data}
\name{PolyhedronScraperConfiguration.class}
\alias{PolyhedronScraperConfiguration.class}
\title{PolyhedronScraperConfiguration}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
PolyhedronScraperConfiguration.class
}
\description{
Abstract class for configuring specific scrapers for Diferent sources
}
\section{Methods}{

\describe{
  \item{\code{initialize()}}{Initializes the object}
}
}

\keyword{datasets}
