% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db-lib.R
\docType{class}
\name{PolyhedronScraperConfigurationDmccoey.class}
\alias{PolyhedronScraperConfigurationDmccoey.class}
\title{PolyhedronScraperConfigurationDmccoey}
\format{\code{\link{R6Class}} object.}
\usage{
PolyhedronScraperConfigurationDmccoey.class
}
\description{
Scraper configuration for Dmccoey source
}
\section{Methods}{

\describe{
  \item{\code{initialize()}}{initializes the object}
  \item{\code{getPolyhedraFiles(home.dir.data)}}{returns the file names on the netlib database}
  \item{\code{scrape(polyhedron.file.id, source.filename)}}{scrapes the object}
}
}

\keyword{datasets}
