% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db-lib.R
\docType{class}
\name{PolyhedronScraperConfigurationNetlib.class}
\alias{PolyhedronScraperConfigurationNetlib.class}
\title{PolyhedronScraperConfigurationNetlib}
\format{\code{\link{R6Class}} object.}
\usage{
PolyhedronScraperConfigurationNetlib.class
}
\description{
Scraper configuration for Netlib source
}
\section{Methods}{

\describe{
  \item{\code{initialize()}}{initializes the object}
  \item{\code{getPolyhedraFiles(home.dir.data)}}{returns the file names on the netlib database}
  \item{\code{scrape(polyhedron.number, polyhedron.filename)}}{scrapes the object}
}
}

\keyword{datasets}
