% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polyhedra-lib.R
\name{PolyhedronState.class}
\alias{PolyhedronState.class}
\title{Polyhedron State}
\arguments{
\item{source}{the source file}

\item{file.id}{the file id}

\item{current.error}{the error to add}

\item{transformation.matrix}{the transformation matrix to apply to the polyhedron}
}
\value{
A new  PolyhedronState object.
'@description
Adds an error to the error string and log it as info

string with polyhedron name
}
\description{
Create a polyhedronState object

Scrapes the polyhedra folder files

Get Polyhedron name

Returns the object corresponding to the solid

Checks edge consistency

Apply transformation matrix to polyhedron

Creates a 'rgl' representation of the object

Gets an XML representation out of the polyhedron object
}
\details{
This abstract class provide the basis from which every polyhedron state class derivate.
}
\section{Fields}{

\describe{
\item{\code{source}}{polyhedron definition source}

\item{\code{file.id}}{polyhedron file id}

\item{\code{errors}}{Errors string}
}}

