% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polyhedra-lib.R
\docType{class}
\name{PolyhedronStateDeserializer.class}
\alias{PolyhedronStateDeserializer.class}
\title{Polyhedron State Deserializer}
\arguments{
\item{serialized.polyhedron}{a serialized polyhedron}
}
\value{
A new  PolyhedronStateDeserializer object.

A new  PolyhedronStateDefined object.
}
\description{
Polyhedron state for deserialize from database

Initialize PolyhedronStateDeserializer object

Generates a PolyhedronStateDefined from a serialized polyhedron
}
\section{Fields}{

\describe{
\item{\code{serialized.polyhedron}}{polyhedron definition serialized}
}}

