% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polyhedra-lib.R
\docType{class}
\name{PolyhedronStateDeserializer}
\alias{PolyhedronStateDeserializer}
\title{PolyhedronStateDeserializer}
\description{
Polyhedron state for deserialize from database
}
\author{
ken4rab
}
\section{Super class}{
\code{\link[Rpolyhedra:PolyhedronState]{Rpolyhedra::PolyhedronState}} -> \code{PolyhedronStateDeserializer}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{serialized.polyhedron}}{polyhedron definition serialized}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-PolyhedronStateDeserializer-new}{\code{PolyhedronStateDeserializer$new()}}
\item \href{#method-PolyhedronStateDeserializer-scrape}{\code{PolyhedronStateDeserializer$scrape()}}
\item \href{#method-PolyhedronStateDeserializer-clone}{\code{PolyhedronStateDeserializer$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="Rpolyhedra" data-topic="PolyhedronState" data-id="addError"><a href='../../Rpolyhedra/html/PolyhedronState.html#method-PolyhedronState-addError'><code>Rpolyhedra::PolyhedronState$addError()</code></a></li>
<li><span class="pkg-link" data-pkg="Rpolyhedra" data-topic="PolyhedronState" data-id="applyTransformationMatrix"><a href='../../Rpolyhedra/html/PolyhedronState.html#method-PolyhedronState-applyTransformationMatrix'><code>Rpolyhedra::PolyhedronState$applyTransformationMatrix()</code></a></li>
<li><span class="pkg-link" data-pkg="Rpolyhedra" data-topic="PolyhedronState" data-id="buildRGL"><a href='../../Rpolyhedra/html/PolyhedronState.html#method-PolyhedronState-buildRGL'><code>Rpolyhedra::PolyhedronState$buildRGL()</code></a></li>
<li><span class="pkg-link" data-pkg="Rpolyhedra" data-topic="PolyhedronState" data-id="checkEdgesConsistency"><a href='../../Rpolyhedra/html/PolyhedronState.html#method-PolyhedronState-checkEdgesConsistency'><code>Rpolyhedra::PolyhedronState$checkEdgesConsistency()</code></a></li>
<li><span class="pkg-link" data-pkg="Rpolyhedra" data-topic="PolyhedronState" data-id="exportToXML"><a href='../../Rpolyhedra/html/PolyhedronState.html#method-PolyhedronState-exportToXML'><code>Rpolyhedra::PolyhedronState$exportToXML()</code></a></li>
<li><span class="pkg-link" data-pkg="Rpolyhedra" data-topic="PolyhedronState" data-id="getName"><a href='../../Rpolyhedra/html/PolyhedronState.html#method-PolyhedronState-getName'><code>Rpolyhedra::PolyhedronState$getName()</code></a></li>
<li><span class="pkg-link" data-pkg="Rpolyhedra" data-topic="PolyhedronState" data-id="getSolid"><a href='../../Rpolyhedra/html/PolyhedronState.html#method-PolyhedronState-getSolid'><code>Rpolyhedra::PolyhedronState$getSolid()</code></a></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PolyhedronStateDeserializer-new"></a>}}
\if{latex}{\out{\hypertarget{method-PolyhedronStateDeserializer-new}{}}}
\subsection{Method \code{new()}}{
Initialize PolyhedronStateDeserializer object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PolyhedronStateDeserializer$new(serialized.polyhedron)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{serialized.polyhedron}}{a serialized polyhedron}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new  PolyhedronStateDeserializer object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PolyhedronStateDeserializer-scrape"></a>}}
\if{latex}{\out{\hypertarget{method-PolyhedronStateDeserializer-scrape}{}}}
\subsection{Method \code{scrape()}}{
Generates a PolyhedronStateDefined from a serialized polyhedron
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PolyhedronStateDeserializer$scrape()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A new  PolyhedronStateDefined object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PolyhedronStateDeserializer-clone"></a>}}
\if{latex}{\out{\hypertarget{method-PolyhedronStateDeserializer-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PolyhedronStateDeserializer$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
