% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polyhedra-lib.R
\docType{class}
\name{PolyhedronStateDmccoeyScraper.class}
\alias{PolyhedronStateDmccoeyScraper.class}
\title{Polyhedron State Dmccoey Scraper}
\format{\code{\link{R6Class}} object.}
\usage{
PolyhedronStateDmccoeyScraper.class
}
\description{
Scrapes polyhedra from a dmccooey file format
}
\section{Methods}{

\describe{
  \item{\code{initialize(number, netlib.p3.lines)}}{Initializes the object, taking the number and PDH file as parameters}
  \item{\code{scrape()}}{Scrapes data from dmccooey file format}
  \item{\code{scrapeValues(values.lines)}}{Scrapes values}
  \item{\code{scrapeVertices(vertices.lines)}}{Scrapes vertices}
  \item{\code{scrapeFaces(face.lines)}}{Scrapes faces}
  \item{\code{applyTransformationMatrix(transformation.matrix)}}{Apply transformation matrix to polyhedron}
  \item{\code{buildRGL(transformation.matrix)}}{Builds the \code{RGL} implementation}
}
}

\keyword{datasets}
