% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polyhedra-lib.R
\docType{class}
\name{PolyhedronStateDmccooeyScraper.class}
\alias{PolyhedronStateDmccooeyScraper.class}
\title{Polyhedron State Dmccooey Scraper}
\arguments{
\item{file.id}{identifier of the definition file.}

\item{polyhedra.dmccooey.lines}{raw Dmccooey definition file lines}

\item{values.lines}{values definitions in Dmccooey source}

\item{vertices.lines}{vertices definitions in Dmccooey source}

\item{faces.lines}{face}

\item{transformation.matrix}{the transformation matrix to apply to the polyhedron}
}
\value{
A new  PolyhedronStateDmccooeyScraper object.

This PolyhedronStateDmccooeyScraper object with regexp defined.

This PolyhedronStateDmccooeyScraper object with values defined.

This PolyhedronStateDmccooeyScraper object with faces defined.

This PolyhedronStateDmccooeyScraper object with faces defined.

A new PolyhedronStateDefined object.

string with polyhedron name
}
\description{
Initialize Dmccooey scraper

setupRegexp for Dmccooey definition

scrape values from Dmccooey definition

scrape polyhedron vertices from definition

scrape polyhedron faces from definition

scrape Dmccooey polyhedron definition

get Polyhedron name

Apply transformation matrix to polyhedron

Creates a 'rgl' representation of the object

serializes object in XML
}
\details{
Scrapes polyhedra from a dmccooey file format
}
\section{Fields}{

\describe{
\item{\code{regexp.values.names}}{regexp for scraping values names}

\item{\code{regexp.rn}}{regexp for scraping real numbers}

\item{\code{regexp.values}}{regexp for scraping values}

\item{\code{regexp.vertex}}{regexp for scraping vertices}

\item{\code{regexp.faces}}{regexp for scraping faces}

\item{\code{polyhedra.dmccooey.lines}}{dmccooey polyhedra definition lines}

\item{\code{labels.map}}{labels map where values are}

\item{\code{values}}{labels map where values are}

\item{\code{vertices}}{specification}

\item{\code{vertices.replaced}}{3D values}

\item{\code{faces}}{definition}
}}

