% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polyhedra-lib.R
\name{PolyhedronStateNetlibScraper.class}
\alias{PolyhedronStateNetlibScraper.class}
\title{Polyhedron State Netlib Scraper}
\arguments{
\item{file.id}{the file id}

\item{netlib.p3.lines}{the lines to add}

\item{label.number}{the label number}

\item{expected.label}{the expected label}

\item{net.txt}{a vector containing net model in netlib format}

\item{offset}{in numbering vertices}

\item{x}{a string containing brackets}

\item{vertices.txt}{vector containing netlib format vertices}

\item{label}{the label to get data from}

\item{transformation.matrix}{the transformation matrix to apply to the polyhedron}
}
\value{
A new  PolyhedronStateNetlibScraper object.

a list containing labels from netlib file description

a list containing a net model

value

data.frame containing netlib vertices

data.frame containing netlib vertices

value

string with polyhedron name

A new PolyhedronStateDefined object.
}
\description{
Initializes the object, taking the file.id and PDH file as parameters

Extracts data from the label, taking the label number and the
  expected label as parameters

get Labels from current netlib file description

scrape Net Model from netlib format

Remove brackets for current field content

scrape vertices described in netlib format

setupLabelsOrder

Get data from label specified as parameter

get Polyhedron name

scrape Netlib polyhedron definition

Apply transformation matrix to polyhedron

Creates a 'rgl' representation of the object

serializes object in XML
}
\details{
Scrapes polyhedra from a PHD file format.
}
\section{Fields}{

\describe{
\item{\code{netlib.p3.lines}}{The path to the PHD files}

\item{\code{labels.rows}}{Labels - row of appearance}

\item{\code{labels.map}}{Labels - Map of content}

\item{\code{errors}}{the errors found}
}}

