% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polyhedra-lib.R
\docType{data}
\name{PolyhedronStateScraper.class}
\alias{PolyhedronStateScraper.class}
\title{Scrapes each polyhedra from PHD file}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
PolyhedronStateScraper.class
}
\description{
Scrapes each polyhedra from PHD file
}
\keyword{datasets}
