% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db-lib.R
\docType{class}
\name{PolyhedronTestTask.class}
\alias{PolyhedronTestTask.class}
\title{PolyhedronTestTask}
\format{\code{\link{R6Class}} object.}
\usage{
PolyhedronTestTask.class
}
\description{
Is an abstract class for specifying TestTask to make R6 iteration methods
like cover complaint with testhat infrastructure
}
\section{Methods}{

\describe{
  \item{\code{initialize()}}{initializes the object}
  \item{\code{run()}}{Run the test task}
}
}

\keyword{datasets}
