% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db-lib.R
\docType{class}
\name{PolyhedronTestTaskScrape.class}
\alias{PolyhedronTestTaskScrape.class}
\title{PolyhedronTestTaskScrape}
\format{\code{\link{R6Class}} object.}
\usage{
PolyhedronTestTaskScrape.class
}
\description{
A Test task for comparing a new scrape with an already scraped polyhedron
in the database
}
\section{Methods}{

\describe{
  \item{\code{initialize()}}{initializes the object}
  \item{\code{run()}}{Run the test task}
}
}

\keyword{datasets}
