% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rpolyhedra-package.R
\docType{package}
\name{Rpolyhedra-package}
\alias{Rpolyhedra-package}
\alias{_PACKAGE}
\alias{Rpolyhedra}
\title{Rpolyhedra: Polyhedra Database}
\description{
A polyhedra database scraped from various sources as R6 objects and 'rgl' visualizing capabilities.
}
\details{
A polyhedra database scraped from:
\itemize{
   \item http://paulbourke.net/dataformats/phd/: PHD files as R6 objects and 'rgl'
         visualizing capabilities. The PHD format was created to describe the geometric
         polyhedra definitions derived mathematically <http://www.netlib.org/polyhedra/>
         by Andrew Hume and by the Kaleido program of Zvi Har'El.
   \item http://dmccooey.com/Polyhedra: Polyhedra text datafiles.
}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/qbotics/Rpolyhedra}
  \item Report bugs at \url{https://github.com/qbotics/Rpolyhedra/issues}
}

}
\author{
\strong{Maintainer}: Alejandro Baranek \email{abaranek@dc.uba.ar} [compiler, copyright holder]

Authors:
\itemize{
  \item Leonardo Belen \email{leobelen@gmail.com} [compiler, copyright holder]
}

}
