% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{rpolyhedra}
\alias{rpolyhedra}
\alias{.onLoad}
\alias{rpolyhedra-package}
\title{rpolyhedra}
\usage{
.onLoad(libname, pkgname)
}
\arguments{
\item{libname}{Library name}

\item{pkgname}{Package name}
}
\description{
A package to replace the now defunct poly command.
}
\author{
Alejandro Baranek <alejandrobaranek@gmail.com>, Leonardo Javier Belen <leobelen@gmail.com>
Executes code while loading the package.
}
