% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polyhedra-lib.R
\name{scrapePolyhedra}
\alias{scrapePolyhedra}
\title{scrape all polyhedra in data dir and saves a representation.}
\usage{
scrapePolyhedra(max.quant = 0, test = TRUE, save.rds.force = FALSE)
}
\arguments{
\item{max.quant}{maximum quantity of polyhedra to scrape}

\item{test}{scrape polyhedra and compare it with preloaded version}

\item{save.rds.force}{force save of database scraped}
}
\description{
scrape all polyhedra in data dir and saves a representation.
}
\examples{
## Scrape first 15 polyhedra in directory
scrapePolyhedra(max.quant = 15, test = FALSE)


}
