% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db-lib.R
\docType{data}
\name{ScraperRecord.class}
\alias{ScraperRecord.class}
\title{ScraperRecord
TODO add doc}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
ScraperRecord.class
}
\description{
ScraperRecord
TODO add doc
}
\keyword{datasets}
