% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/env-lib.R
\name{selectDataEnv}
\alias{selectDataEnv}
\title{selectDataEnv}
\usage{
selectDataEnv(env=NA, prompt.value = NULL)
}
\arguments{
\item{env}{The environment to run on, can be PACKAGE, HOME or NULL. If null, it asks the user for a an Environment.}

\item{prompt.value}{If specified, no prompt is shown}
}
\value{
.data.env
}
\description{
Asks the user where to set the system variable .data.env
}
