% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db-lib.R
\name{setPackageEnvir}
\alias{setPackageEnvir}
\title{setPackageEnvir}
\usage{
setPackageEnvir(variable.name, value)
}
\arguments{
\item{variable.name}{name of variable to be set}

\item{value}{variable value}
}
\description{
Set a variable from package environment
}
