% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db-lib.R
\name{setUserEnvir}
\alias{setUserEnvir}
\title{setUserEnvir}
\usage{
setUserEnvir(variable.name, value)
}
\arguments{
\item{variable.name}{name of variable to be set}

\item{value}{variable value}
}
\description{
Sets variable in the writable environment for the package
}
