% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db-lib.R
\name{switchToFullDatabase}
\alias{switchToFullDatabase}
\title{switchToFullDatabase()}
\usage{
switchToFullDatabase(env=NA)
}
\arguments{
\item{env}{The environment to run on, can be PACKAGE, HOME or NA. If NA, it asks the user for a an Environment.}
}
\value{
.data.env
}
\description{
Prompts user for changing database to fulldb in user filespace
}
