% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db-lib.R
\name{updatePolyhedraDatabase}
\alias{updatePolyhedraDatabase}
\title{updatePolyhedraDatabase}
\usage{
updatePolyhedraDatabase(source.filenames = NULL)
}
\arguments{
\item{source.filenames}{if not null specify which source filenames to scrape}
}
\value{
polyhedra db object
}
\description{
Function for initializing database
}
