% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml-lib.R
\name{validatePolyhedronXML}
\alias{validatePolyhedronXML}
\title{validatePolyhedronXML()}
\usage{
validatePolyhedronXML(polyhedron.xml)
}
\arguments{
\item{polyhedron.xml}{the xml document object to test against the schema}
}
\value{
the document or an error status.
}
\description{
Validates the xml against the schema.
}
