% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db-lib.R
\name{copyFilesToExtData}
\alias{copyFilesToExtData}
\title{downloadRPolyhedraSupportingFiles}
\usage{
copyFilesToExtData(force = FALSE)
}
\value{
TRUE if sucessfull, FALSE otherwise
}
\description{
Downloads the files from the remote location
}
