% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db-lib.R
\name{copyFilesToExtData}
\alias{copyFilesToExtData}
\title{downloadRPolyhedraSupportingFiles}
\usage{
copyFilesToExtData(force = FALSE)
}
\arguments{
\item{force}{indicate if existings directories must be overwritten}
}
\value{
TRUE if sucessfull,
}
\description{
Downloads the files from the remote location
}
