% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db-lib.R
\name{copyFilesToExtData}
\alias{copyFilesToExtData}
\title{copyFilesToExtData}
\usage{
copyFilesToExtData(force = FALSE)
}
\arguments{
\item{force}{indicate if existings directories must be overwritten}
}
\value{
TRUE if sucessfull
}
\description{
Copies files from the home directory to the package one to build a preconfigured package.
}
