% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/public-lib.R
\name{getAvailableSources}
\alias{getAvailableSources}
\title{Get available sources}
\usage{
getAvailableSources()
}
\value{
sources string vector, which can be obtained from getAvailableSources()
}
\description{
Gets the list of names of available sources in database to be used later as
references to the package.
}
\examples{
# gets all sources in the database
available.sources <- getAvailableSources()

# returns all polyhedra from all sources
available.polyhedra <- getAvailablePolyhedra(sources = available.sources)

# search within the polyhedron names from all sources
cubes <- getAvailablePolyhedra(
  sources = available.sources,
  search.string = "cube"
)
cubes
}
\seealso{
getAvailablePolyhedra, getPolyhedron
}
