% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db-lib.R
\name{getDataDir}
\alias{getDataDir}
\title{getDataDir}
\usage{
getDataDir(data.env = getDataEnv())
}
\arguments{
\item{data.env}{enviroment where data directory must be returned}
}
\value{
dir where the package access polyhedra database
}
\description{
Gets the path of Rpolyhedra data dir.
}
\details{
This function is used internally to determine whether the package
is compiled in source or package directory.
}
