% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db-lib.R
\name{getDataDir}
\alias{getDataDir}
\title{Gets the path of package data.}
\usage{
getDataDir()
}
\description{
This function is used internally to determine whether the package
is compiled in source or package directory.
}
