% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db-lib.R
\name{getPackageEnvir}
\alias{getPackageEnvir}
\title{getPackageEnvir}
\usage{
getPackageEnvir(variable.name)
}
\arguments{
\item{variable.name}{name of variable to be retrieved
Obtains a variable from package environment}
}
\description{
getPackageEnvir
}
