% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db-lib.R
\name{getPercentilPolyhedraQuant}
\alias{getPercentilPolyhedraQuant}
\title{getPercentilPolyhedraQuant
returns polyhedra quantity of parameter percentil}
\usage{
getPercentilPolyhedraQuant(percentil)
}
\arguments{
\item{percentil}{is the percentil which must be applied to estimate the figure}
}
\description{
getPercentilPolyhedraQuant
returns polyhedra quantity of parameter percentil
}
