% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db-lib.R
\name{getPercentilPolyhedraQuant}
\alias{getPercentilPolyhedraQuant}
\title{getPercentilPolyhedraQuant
returns polyhedra quantity of parameter percentil}
\usage{
getPercentilPolyhedraQuant(percentil, quant.min = 100)
}
\arguments{
\item{percentil}{is the percentil which must be applied
to estimate the figure}

\item{quant.min}{minimum quantity of files to return}
}
\description{
getPercentilPolyhedraQuant
returns polyhedra quantity of parameter percentil
}
