% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db-lib.R
\name{getPolyhedraRDSPath}
\alias{getPolyhedraRDSPath}
\title{Gets the path of Polyhedra RDS database file}
\usage{
getPolyhedraRDSPath(polyhedra_rds_filename = "polyhedra.RDS")
}
\arguments{
\item{polyhedra_rds_filename}{filename of polyhedra database}
}
\value{
the path to the Polyhedra database file
}
\description{
Gets the path of Polyhedra RDS database file
}
