% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db-lib.R
\name{getUserEnvir}
\alias{getUserEnvir}
\title{getUserEnvir}
\usage{
getUserEnvir(variable.name)
}
\arguments{
\item{variable.name}{name of variable to be retrieved}
}
\description{
Gets a writable environment for the package
}
