% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serialization-lib.R
\name{hydratePolyhedron}
\alias{hydratePolyhedron}
\title{hydratePolyhedron}
\usage{
hydratePolyhedron(file.path)
}
\arguments{
\item{file.path}{the path of the file to be hydrated.}
}
\value{
polyhedron.state.defined
}
\description{
hydrates a polyhedron from the defined path
}
