% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db-lib.R
\name{isCompatiblePolyhedraRDS}
\alias{isCompatiblePolyhedraRDS}
\title{isCompatiblePolyhedraRDS()}
\usage{
isCompatiblePolyhedraRDS(.polyhedra.candidate = getPolyhedraObject(),
  halts = FALSE)
}
\arguments{
\item{.polyhedra.candidate}{polyhedra db to test}

\item{halts}{indicates whether it has to halt execution when it is not compatible}
}
\description{
Tests if the polyhedra RDS is compatible with the current format
}
