% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/env-lib.R
\name{loggerSetupFile}
\alias{loggerSetupFile}
\title{loggerSetupFile}
\usage{
loggerSetupFile(log.file)
}
\arguments{
\item{log.file}{filepath to configure logger AppenderFile}
}
\description{
loggerSetupFile
}
\author{
ken4rab
}
