% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polyhedra-lib.R
\docType{class}
\name{Polyhedron.class}
\alias{Polyhedron.class}
\title{Polyhedron}
\arguments{
\item{state}{polyhedron state object}

\item{file.id}{the file id}

\item{netlib.p3.lines}{vector with netlib definition lines}

\item{polyhedra.dmccooey.lines}{vector with Dmccooey definition lines}

\item{serialized.polyhedron}{a serialized version of a polyhedron state}

\item{transformation.matrix}{transformation matrix parameter}

\item{expected.vertices}{expected vertices number}

\item{expected.faces}{expected faces number}
}
\value{
A new  Polyhedron object.

A new  PolyhedronStateDefined object.

A new  PolyhedronStateDefined object.

A new  PolyhedronStateDefined object.

string with polyhedron name

A new  PolyhedronState object.

A list of vertex vectors composing polyhedron faces.

A boolean value

An tmesh3d object

A character object with the XML definition

a data.frame with polyhedron errors

Unmodified polyhedron object
}
\description{
Polyhedron container class, which is accessible by the final users upon call

Create a polyhedronState object

scrape Netlib polyhedron definition

scrape Dmccooey polyhedron definition

deserialize a polyhedron state definition

get Polyhedron name

Gets polyhedron state

Gets a solid definition

checks Edges consistency

Return an 'rgl' model with an optional transformation described by transformation.matrix parameter

exports an XML definition of current polyhedron

returns the errors found when processing current polyhedron

check properties of current polyhedron
}
\section{Fields}{

\describe{
\item{\code{file.id}}{Polyhedron file.id}

\item{\code{state}}{Polyhedron state}
}}

