% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polyhedra-lib.R
\docType{data}
\name{Polyhedron.class}
\alias{Polyhedron.class}
\title{Polyhedron container class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Polyhedron.class
}
\description{
Polyhedron container class
}
\keyword{datasets}
