% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polyhedra-lib.R
\docType{class}
\name{Polyhedron.class}
\alias{Polyhedron.class}
\title{Polyhedron}
\format{\code{\link{R6Class}} object.}
\usage{
Polyhedron.class
}
\description{
Polyhedron container class, which is accesible by the final users upon call to \code{getPolyhedron()}
}
\section{Fields}{

\describe{
\item{\code{number}}{Polyhedron number}

\item{\code{state}}{Polyhedron state}
}}

\section{Methods}{

\describe{
  \item{\code{initialize(number, state = NULL)}}{Initializes the object}
  \item{\code{scrapeNetlib(polyhedron.lines)}}{{Scrapes polyhedra from the definition}}
  \item{\code{getName()}}{Gets the name from polyhedron definition}
  \item{\code{getState()}}{Gets the state from polyhedron definition}
  \item{\code{getSolid()}}{Gets the solid definition of polyhedron definition}
  \item{\code{isChecked()}}{Returns TRUE is polyhedron is checked}
  \item{\code{getErrors()}}{Returns errors collected in checking process}
  \item{\code{applyTransformationMatrix(transformation.matrix)}}{Apply transformation matrix to polyhedron}
  \item{\code{getRGLModel(size = 1, origin = c(0, 0, 0)}}{Builds the RGL model}
  \item{\code{exportToXML()}}{Gets an XML representation out of the polyhedron object}
  \item{\code{checkProperties(expected.vertices, expected.faces)}}{check polyhedron basic properties}

}
}

\keyword{datasets}
