% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polyhedra-lib.R
\docType{data}
\name{PolyhedronStateDefined.class}
\alias{PolyhedronStateDefined.class}
\title{PolyhedronState when is it defined. Processes the definitionin order to produce a graphical representation in RGL.}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
PolyhedronStateDefined.class
}
\description{
PolyhedronState when is it defined. Processes the definitionin order to produce a graphical representation in RGL.
}
\keyword{datasets}
