% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serialization-lib.R
\name{polyhedronToXML}
\alias{polyhedronToXML}
\title{polyhedronToXML()}
\usage{
polyhedronToXML(polyhedron.state.defined)
}
\arguments{
\item{polyhedron.state.defined}{the polyhedron to get a representation from}
}
\value{
a XML document, ready to be saved as a String File
}
\description{
Gets an XML representation out of the polyhedron object
}
\examples{
#get the representation of a cube (netlib library)
library(Rpolyhedra)
polyhedronToXML(getPolyhedron("netlib", "cube")$getState())

}
