% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serialization-lib.R
\name{polyhedronToXML}
\alias{polyhedronToXML}
\title{Polyhedron to XML}
\usage{
polyhedronToXML(polyhedron.state.defined, is.transformed.vertices = TRUE)
}
\arguments{
\item{polyhedron.state.defined}{the polyhedron to get a representation from}

\item{is.transformed.vertices}{flag which states if vertices are in original position or transformationMatrix applied}
}
\value{
an XML document, ready to be converted to String with XML::saveXML()
}
\description{
Gets an XML representation out of the polyhedron object
}
\examples{
# get the representation of a cube (netlib library)
XML::saveXML(polyhedronToXML(getPolyhedron("netlib", "cube")$state))

}
