% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rpolyhedra-package.R
\docType{package}
\name{Rpolyhedra-package}
\alias{Rpolyhedra-package}
\alias{Rpolyhedra}
\title{Rpolyhedra: Polyhedra Database}
\description{
A polyhedra database scraped from various sources as R6 objects and 'rgl' visualizing capabilities.
}
\details{
A polyhedra database scraped from:
\itemize{
   \item http://paulbourke.net/dataformats/phd/: PHD files as R6 objects and 'rgl'
         visualizing capabilities. The PHD format was created to describe the geometric
         polyhedra definitions derived mathematically <https://netlib.org/polyhedra/>
         by Andrew Hume and by the Kaleido program of Zvi Har'El.
   \item http://dmccooey.com/Polyhedra/: Polyhedra text datafiles.
}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://docs.ropensci.org/Rpolyhedra/}
  \item \url{https://github.com/ropensci/Rpolyhedra}
  \item Report bugs at \url{https://github.com/ropensci/Rpolyhedra/issues}
}

}
\author{
\strong{Maintainer}: Alejandro Baranek \email{abaranek@dc.uba.ar} [compiler, copyright holder]

Authors:
\itemize{
  \item Leonardo Belen \email{leobelen@gmail.com} [compiler, copyright holder]
}

Other contributors:
\itemize{
  \item  qbotics \email{qbotics6@gmail.com} [copyright holder]
  \item Barret Schloerke \email{schloerke@gmail.com} [reviewer]
  \item Lijia Yu \email{yu@lijiayu.net} [reviewer]
}

}
