% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/public-lib.R
\name{scrapePolyhedra}
\alias{scrapePolyhedra}
\title{Scrape polyhedra objects}
\usage{
scrapePolyhedra(
  scrape.config,
  source.filenames = NULL,
  sources.config = getUserEnvir(".available.sources"),
  logger = lgr
)
}
\arguments{
\item{scrape.config}{predefined configuration for scraping}

\item{source.filenames}{if not null specify which source filenames to scrape}

\item{sources.config}{the sources that will be used by the function}

\item{logger}{logger for inheriting threshold from calling class/function}
}
\value{
polyhedra db object
}
\description{
Gets polyhedra objects from text files of
different sources, scheduling and scraping using
predefined configurations.
}
