% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db-lib.R
\name{scrapePolyhedra}
\alias{scrapePolyhedra}
\title{scrapePolyhedra()}
\usage{
scrapePolyhedra(scrape.config, sources.config = get(".available.sources",
  envir = asNamespace("Rpolyhedra")))
}
\arguments{
\item{scrape.config}{predefined configuration for scraping}

\item{sources.config}{the sources that will be used by the function}
}
\description{
Gets polyhedra objects from text files of
different sources, scheduling and scraping using predefined configurations
}
