% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db-lib.R
\name{selectDataEnv}
\alias{selectDataEnv}
\title{selectDataEnv}
\usage{
selectDataEnv(env=NA)
}
\arguments{
\item{env}{The environment to run on, can be PACKAGE, HOME or NULL. If null, it asks the user for a an Environment.}
}
\value{
.data.env
}
\description{
Asks the user where to set the system variable .data.env
}
